Safari-611.1.10/LayoutTests/inspector/animation/lifecycle-web-animation-expected.txt

Tests for the Animation.animationCreated and Animation.animationDestroyed events.


== Running test suite: Animation.Lifecycle
-- Running test case: Animation.Lifecycle.WebAnimation
PASS: There should not be any animations.
Creating animation...

PASS: Animation created 'web-animation-test'.
  0: animate - [native code]
  1: createAnimation - inspector/animation/resources/lifecycle-utilities.js:7:35
  2: Global Code - [program code]
  3: evaluateWithScopeExtension - [native code]
  4: (anonymous function) - [native code]
  5: _wrapCall - [native code]
PASS: Animation type should be Web Animation.
startDelay: 100
endDelay: 200
iterationCount: 10
iterationStart: 5
iterationDuration: 300
timingFunction: "ease-in-out"
playbackDirection: "alternate"
fillMode: "both"
keyframes:
[
  {
    "offset": 0.25,
    "easing": "cubic-bezier(0.1, 0.2, 0.3, 0.4)",
    "style": "color: red;\nopacity: 0;"
  },
  {
    "offset": 0.75,
    "easing": "cubic-bezier(0.6, 0.7, 0.8, 0.9)",
    "style": "color: blue;\nopacity: 1;"
  }
]
